+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.1.9 -- Developing...
 * DataGridSample is more compact and clear now.
 * DatePicker bugs fixed.
-* Button component added for issue #12.
+* Button component added for issue #12.*
+* ModalDataForm uses new Button for submit.
 
 ## 1.1.8
 * robe-react-commons updated